+++ test/moengage/moengage_service_test.mocks.dart
@@ -31,11 +31,18 @@ class MockMoEngagePlatform extends _i1.Mock implements _i2.MoEngagePlatform {
   }
 
   @override
-  void initialize({required String? appId}) => super.noSuchMethod(
+  void initialize({
+    required String? appId,
+    String? pushToken,
+  }) =>
+      super.noSuchMethod(
         Invocation.method(
           #initialize,
           [],
-          {#appId: appId},
+          {
+            #appId: appId,
+            #pushToken: pushToken,
+          },
         ),
         returnValueForMissingStub: null,
       );